+++ packages/react/src/experimental/exports.ts
@@ -16,6 +16,7 @@ export * from "./OneDataCollection/exports"
 export * from "./OnePagination"
 export * from "./Overlays/exports"
 export * from "./RichText/exports"
+export * from "./UpsellingKit/exports"
 export * from "./Utilities/exports"
 export * from "./Widgets/exports"